002 Add Two Numbers: Remove floating point errors (#56)
Contributed by @dannysepler

+++ python/002_Add_Two_Numbers.py
@@ -51,7 +51,7 @@ class Solution(object):
                 l2 = l2.next
             curr.next = ListNode(val % 10)
             curr = curr.next
-            carry = val / 10
+            carry = int(val / 10)
         if carry > 0:
             curr.next = ListNode(carry)
         return head.next
